backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
# LLM Integrations - choose one or more
openai==1.3.0          # GPT, GPT-4 (реально работающий пакет)
# anthropic==0.7.0      # Claude API (опционально)
# huggingface-hub==0.17.0  # Hugging Face models (опционально)
